refactor and improve simulation
