+++ dynestyx/simulation/__init__.py
@@ -1,7 +1,7 @@
 """Simulation backends and simulator handlers."""
 
 from dynestyx.inference.utils.plate_utils import _slice_tree_for_plate_member
-from dynestyx.simulation.base import Simulator
+from dynestyx.simulation.auto import Simulator
 from dynestyx.simulation.discrete import DiscreteTimeSimulator
 from dynestyx.simulation.ode import ODESimulator
 from dynestyx.simulation.sde import SDESimulator
